[Refactor] Category를 ProductManage 페이지에서 관리하도록 코드를 이전함
