+++ src/main/webapp/resources/css/product_manage/productManage.css
@@ -19,6 +19,7 @@ h2 {
     font-weight: bold;
     text-align: center;
     margin : 30px;
+    padding-top: 50px;
 }
 
 button {
@@ -86,8 +87,8 @@ button {
 
 .head {
     display: block;
-    border-top: 2px solid red;
-    border-bottom: 2px solid red;
+    border-top: 2px solid black;
+    border-bottom: 2px solid black;
     overflow: hidden;
     width: 100%;
 }
@@ -217,8 +218,8 @@ button {
 .product-options #product-option-form .option-head {
     overflow: hidden;
     width: 100%;
-    border-top: 2px solid green;
-    border-bottom: 2px solid green;
+    border-top: 2px solid black;
+    border-bottom: 2px solid black;
     padding-top: 10px;
     padding-bottom: 10px;
 }
@@ -305,8 +306,8 @@ button {
 .product-options #product-option-detail-form .option-detail-head {
     overflow: hidden;
     width: 100%;
-    border-top: 2px solid blue;
-    border-bottom: 2px solid blue;
+    border-top: 2px solid black;
+    border-bottom: 2px solid black;
     padding-top: 10px;
     padding-bottom: 10px;
     margin-bottom: 10px;
@@ -404,13 +405,13 @@ button {
 .category .category-wrap {
     width: 40%;
     float: left;
-    border-bottom: 2px solid green;
+    border-bottom: 2px solid black;
 }
 
 .category .category-wrap .category-head {
     overflow: hidden;
-    border-top: 2px solid green;
-    border-bottom: 2px solid green;
+    border-top: 2px solid black;
+    border-bottom: 2px solid black;
     padding: 10px 0;
     text-align: center;
 }
